requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
python-dotenv==1.0.1
mysqlclient==2.2.4
google-cloud-sql-connector==1.5.0 